MAINT: rename to avoid name confusion with existing tools
